Litt mer splitting
